HS-68
Cleaned up the code we used. Streamlined it to be easier to understand and moved everything to its own file. The bot is now fully functional and creates its own machine learning model , retrieves the current price, makes a decision, and then acts on that decision. Tested it by having it run over 3000 seconds for a decision every 300 seconds for a total of 100 decisions.